# openapi.yaml
openapi: 3.0.3
info:
  title: StyliAI API
  version: "1.0.0"
  description: >
    Hand-authored OpenAPI contract for the StyliAI backend, written directly
    from the actual controller/model source (not aspirational) as the single
    source of truth for client-side type generation. Endpoints that don't
    exist yet (e.g. POST /api/styles/preview, called by the dashboard's
    dormant Prompt Preview UI) are intentionally NOT documented here - this
    spec describes what the API actually does today.
servers:
  - url: https://styliai-backend-production.up.railway.app
  - url: http://localhost:3000
components:
  securitySchemes:
    userBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Regular user JWT, signed with SUPABASE_JWT_SECRET.
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT, signed with ADMIN_JWT_SECRET, requires role=admin.
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
      required: [message]
    ErrorCode:
      type: object
      description: Structured error shape used by generateController.js and walletController.js (Roadmap Item 2.1). Other controllers still use ErrorMessage.
      properties:
        code:
          type: string
          example: INSUFFICIENT_BALANCE
        message:
          type: string
      required: [code, message]
...
    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        sortOrder: { type: integer }
        isEnabled: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, sortOrder, isEnabled]
    Style:
      type: object
      description: >
        Admin-shaped Style (includes prompt/negativePrompt/tagIds). Non-admin
        callers (the mobile app) get an allowlisted subset with prompt,
        negativePrompt and tagIds omitted entirely - tags are internal
        ranking metadata, never exposed to end users - see styleModel.getPublicStyles.
      properties:
        id: { type: string, format: uuid }
        categoryId: { type: string, format: uuid }
        name: { type: string }
        prompt: { type: string }
        negativePrompt: { type: string, nullable: true }
        coverImage: { type: string, nullable: true }
        coverImageThumbnail: { type: string, nullable: true, description: "~320x400 WebP browsing thumbnail. Null until the backfill/upload pipeline has generated one." }
        creditCost: { type: integer, minimum: 0 }
        minImages: { type: integer, minimum: 1, default: 1 }
        maxImages: { type: integer, minimum: 1, maximum: 5, default: 1 }
        isTrending: { type: boolean }
        isPremium: { type: boolean }
        isEnabled: { type: boolean }
        sortOrder: { type: integer }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        tagIds:
          type: array
          items: { type: string, format: uuid }
          description: Admin-only. Curated tag ids assigned to this style, used by RecommendationService.
        tagsAutoAssigned:
          type: boolean
          description: Admin-only. True if tagIds was last set by autoTagService rather than a manual dashboard edit - controls whether a future save re-runs classification or preserves a manual override.
        fields:
          type: array
          description: >
            Dynamic input fields backing the prompt's {{placeholder}} tokens.
            Empty for classic styles. Returned on both admin and public reads
            (the field definitions are safe to expose; the prompt is not).
          items: { $ref: '#/components/schemas/StyleField' }
      required: [id, categoryId, name, prompt, negativePrompt, coverImage, creditCost, isTrending, isPremium, isEnabled, sortOrder]
    StyleField:
      type: object
      description: One dynamic input field for a style's prompt template.
      properties:
        key:
          type: string
          description: Placeholder key referenced as {{key}} in the prompt (lower_snake_case).
        label: { type: string }
        type:
          type: string
          enum: [text, textarea, number, dropdown, checkbox, color, date]
          description: Control type. New types may be added without breaking older clients.
        required: { type: boolean }
        placeholder: { type: string, nullable: true }
        options:
          type: array
          nullable: true
          description: For dropdown fields.
          items:
            type: object
            properties:
              value: { type: string }
              label: { type: string }
            required: [value, label]
        config:
          type: object
          additionalProperties: true
          description: Extensible per-type knobs (e.g. min, max, maxLength, default, trueText).
        sortOrder: { type: integer }
      required: [key, label, type, required]
    Tag:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        isEnabled: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, slug, isEnabled]
    WalletInfo:
      type: object
      properties:
        balance: { type: integer }
        adsProgress: { type: integer }
        generatedImages: { type: integer }
        dailyLimitReached: { type: boolean }
      required: [balance, adsProgress, generatedImages, dailyLimitReached]
...
    UploadResult:
      type: object
      properties:
        url: { type: string, format: uri }
        thumbnailUrl:
          type: string
          format: uri
          nullable: true
          description: "~320x400 WebP browsing thumbnail. Null only if thumbnail generation failed - the original upload above still succeeded."
      required: [url, thumbnailUrl]
    GenerateResult:
      type: object
      properties:
        success: { type: boolean }
        generatedImageUrl: { type: string, format: uri }
        thumbnailUrl: { type: string, format: uri, nullable: true }
        generationId:
          type: string
          format: uuid
          nullable: true
          description: "The creations row id for this generation, if the best-effort history write succeeded. Round-trip this back on POST /api/feedback."
        categoryId:
          type: string
          format: uuid
          nullable: true
          description: The generated style's category id, for round-tripping on POST /api/feedback.
        generationTimeMs:
          type: integer
          nullable: true
          description: Server-side AI provider call duration, for round-tripping on POST /api/feedback.
      required: [success, generatedImageUrl, thumbnailUrl]
    GenerationFeedbackSubmission:
      type: object
      description: Post-generation feedback. Never includes image data - only ids/metrics/text.
      properties:
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: string, maxLength: 2000, nullable: true }
        generationId: { type: string, format: uuid, nullable: true }
        styleId: { type: string, format: uuid, nullable: true }
        categoryId: { type: string, format: uuid, nullable: true }
        generationTimeMs: { type: integer, nullable: true }
        appVersion: { type: string, maxLength: 32, nullable: true }
      required: [rating]
    GenerationFeedbackResult:
      type: object
      properties:
        id: { type: string, format: uuid }
        createdAt: { type: string, format: date-time }
      required: [id, createdAt]
...
    AiGenerateResult:
      type: object
      properties:
        success: { type: boolean }
        imageUrl: { type: string, format: uri }
        thumbnailUrl: { type: string, format: uri, nullable: true }
      required: [success, imageUrl, thumbnailUrl]
...
    CreditPack:
      type: object
      description: >
        Roadmap Item 4.3. productId is intentionally nullable - no real
        RevenueCat/App Store/Google Play products exist yet (Item 4.1, not
        started). It must not be treated as a real product identifier until
        that item actually configures one.
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        credits: { type: integer, minimum: 1 }
        priceDisplay: { type: string, example: "$4.99" }
        badge: { type: string, nullable: true }
        description: { type: string, nullable: true }
        isEnabled: { type: boolean }
        sortOrder: { type: integer }
        productId: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, name, credits, priceDisplay, isEnabled, sortOrder, productId]
paths:
...
  /api/categories:
    get:
      summary: List all categories.
      security: []
      responses:
        '200':
          description: Category list
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Category' } }
    post:
      summary: Create a category.
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                isEnabled: { type: boolean, default: true }
              required: [name]
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Category' } } } }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorMessage' } } } }
        '409': { description: Duplicate name, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorMessage' } } } }
...
  /api/styles:
    get:
      summary: >
        List styles. Defaults to enabled-only unless all=true (used by the
        admin dashboard). recommended=true returns the personalized "Recommended
        For You" feed instead (requires userBearerAuth; returns [] if the
        caller is anonymous, personalization is off, or there's no favorite/creation
        history yet - see RecommendationService).
      security: [{ userBearerAuth: [] }, {}]
      parameters:
        - name: categoryId
          in: query
          required: false
          schema: { type: string, format: uuid }
        - name: all
          in: query
          required: false
          schema: { type: string, enum: ["true"] }
        - name: trending
          in: query
          required: false
          schema: { type: string, enum: ["true"] }
        - name: recommended
          in: query
          required: false
          schema: { type: string, enum: ["true"] }
      responses:
        '200':
          description: Style list
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Style' } }
    post:
      summary: Create a style.
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                categoryId: { type: string, format: uuid }
                name: { type: string }
                prompt: { type: string }
                negativePrompt: { type: string, nullable: true }
                coverImage: { type: string, nullable: true }
                coverImageThumbnail: { type: string, nullable: true }
                creditCost: { type: integer, minimum: 0, default: 1 }
                minImages: { type: integer, minimum: 1, default: 1 }
                maxImages: { type: integer, minimum: 1, maximum: 5, default: 1 }
                isTrending: { type: boolean, default: false }
                isPremium: { type: boolean, default: false }
                isEnabled: { type: boolean, default: true }
                sortOrder: { type: integer }
                tagIds:
                  type: array
                  items: { type: string, format: uuid }
                  default: []
                  description: Ignored unless autoAssignTags is explicitly false.
                autoAssignTags:
                  type: boolean
                  default: true
                  description: When true (default), tagIds is computed by autoTagService from name/prompt/category and any client-sent tagIds is ignored. Set false to use tagIds as a manual override.
                fields:
                  type: array
                  description: Dynamic input field definitions. Replaces the style's full field set. Omit to leave unchanged.
                  items: { $ref: '#/components/schemas/StyleField' }
              required: [categoryId, name, prompt]
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Style' } } } }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorMessage' } } } }
        '409': { description: Duplicate name in category, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorMessage' } } } }
...
  /api/tags:
    get:
      summary: List all tags. Admin-only - tags are internal ranking metadata, never listed for the mobile app.
      security: [{ adminBearerAuth: [] }]
      responses:
        '200':
          description: Tag list
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Tag' } }
    post:
      summary: Create a tag.
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                isEnabled: { type: boolean, default: true }
              required: [name]
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/Tag' } } } }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorMessage' } } } }
        '409': { description: Duplicate tag name, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorMessage' } } } }
...
  /api/upload:
    post:
      summary: Upload a cover image to Supabase Storage.
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
              required: [file]
      responses:
        '200': { description: Uploaded, content: { application/json: { schema: { $ref: '#/components/schemas/UploadResult' } } } }
        '400': { description: No file / invalid type, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorMessage' } } } }
    delete:
      summary: Delete a previously-uploaded cover image by its public URL.
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                url: { type: string, format: uri }
              required: [url]
      responses:
        '204': { description: Deleted }
        '400': { description: Missing/invalid URL, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorMessage' } } } }
  /api/generate:
    post:
      summary: Generate a styled image. Charges credits atomically before calling the AI provider, refunds on failure.
      security: [{ userBearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file: { type: string, format: binary }
                styleId: { type: string, format: uuid }
              required: [file, styleId]
      responses:
        '200': { description: Generated, content: { application/json: { schema: { $ref: '#/components/schemas/GenerateResult' } } } }
        '400': { description: Validation error / style disabled, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorCode' } } } }
        '403': { description: Insufficient balance, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorCode' } } } }
        '404': { description: Style not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorCode' } } } }
        '503': { description: AI provider temporarily unavailable, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorCode' } } } }
  /api/feedback:
    post:
      summary: Submits post-generation feedback (star rating + optional comment). Never accepts image data.
      security: [{ userBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/GenerationFeedbackSubmission' }
      responses:
        '201': { description: Feedback recorded, content: { application/json: { schema: { $ref: '#/components/schemas/GenerationFeedbackResult' } } } }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorCode' } } } }
        '401': { description: Missing/invalid user token, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorCode' } } } }
  /api/ai/generate:
    post:
      summary: >
        Generate an image (WebP) from a text prompt via Stability AI. Charges
        credits atomically before calling the AI provider, refunds on failure
        - same flow as /api/generate. Fully independent of /api/generate
        (style-transfer) and the internal auto-tagging service otherwise -
        separate service/controller/route, no shared state. At least one of
        prompt/styleId is required - GET /api/styles never exposes a style's
        prompt text to the client, so styleId lets the caller ask for a
        style's prompt to be resolved server-side instead of supplying it
        directly (mirrors /api/generate's own server-side resolution).
      security: [{ userBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt: { type: string, description: "Raw prompt text. Omit and send styleId instead to have the server resolve it." }
                styleId: { type: string, format: uuid, description: "Resolves prompt/negativePrompt server-side when prompt is omitted." }
                negativePrompt: { type: string }
                aspectRatio: { type: string, example: "16:9" }
                style: { type: string, example: "photographic" }
      responses:
        '200': { description: Generated, content: { application/json: { schema: { $ref: '#/components/schemas/AiGenerateResult' } } } }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorCode' } } } }
        '403': { description: Insufficient balance, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorCode' } } } }
        '429': { description: Rate limited by Stability AI, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorCode' } } } }
        '503': { description: Stability AI unavailable, misconfigured, or out of credits, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorCode' } } } }
  /api/wallet:
    get:
      summary: Get the authenticated user's wallet info.
      security: [{ userBearerAuth: [] }]
      responses:
        '200': { description: Wallet info, content: { application/json: { schema: { $ref: '#/components/schemas/WalletInfo' } } } }
...
  /api/credit-packs:
    get:
      summary: List credit packs. Defaults to enabled-only unless all=true (used by the admin dashboard).
      security: []
      parameters:
        - name: all
          in: query
          required: false
          schema: { type: string, enum: ["true"] }
      responses:
        '200':
          description: Credit pack list
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/CreditPack' } }
    post:
      summary: Create a credit pack.
      security: [{ adminBearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                credits: { type: integer, minimum: 1 }
                priceDisplay: { type: string }
                badge: { type: string, nullable: true }
                description: { type: string, nullable: true }
                isEnabled: { type: boolean, default: true }
                sortOrder: { type: integer }
              required: [name, credits, priceDisplay]
      responses:
        '201': { description: Created, content: { application/json: { schema: { $ref: '#/components/schemas/CreditPack' } } } }
        '400': { description: Validation error, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorMessage' } } } }
        '409': { description: Duplicate name, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorMessage' } } } }
...